refactor: fix clippy warnings across 116 files
- collapsible_if: 786 → 28 (29 nested if blocks collapsed)
- redundant_closure: eliminated (60+ closures simplified)
- unnecessary_cast: eliminated (120+ identity casts removed)
- or_insert_with(T::new) → or_default() where type inference allows
- unused imports cleaned up via cargo fix
- Conflict in cdc/exactly_once.rs resolved (kept safe write pattern)

+++ src/compute/distance_computation/quantized.rs
@@ -1177,7 +1177,7 @@ impl PQDistanceTable {
 
         for (subvector_idx, centroids) in codebook.iter().enumerate() {
             let mut centroid_distances = Vec::with_capacity(num_centroids);
-            let subvector_size = query.len() / num_subvectors as usize;
+            let subvector_size = query.len() / num_subvectors;
             let query_subvector =
                 &query[subvector_idx * subvector_size..(subvector_idx + 1) * subvector_size];
